refactor: consolidate use of EquivalencyOptions
- Always use a Func<EquivalencyOptions, EquivalencyOptions> to adapt the equivalency options
- If not set always use the default value from `Customize.aweXpect.Equivalency().DefaultEquivalencyOptions.Get()`
